timers trigger check avx2

+++ src/mikey/mod.rs
@@ -136,7 +136,7 @@ impl Mikey {
     pub fn tick(&mut self, bus: &mut Bus, cart: &mut Cartridge, dma_ram: &Ram) {
         self.ticks += 1;
 
-        let (mut int, int4_triggered) = self.timers.tick_all(self.ticks);
+        let (mut int, int4_triggered) = self.timers.tick_all(self.ticks as i64);
 
         if int4_triggered {
             // "The interrupt bit for timer 4 (UART baud rate) is driven by receiver or transmitter ready bit of the UART."
